refactor: `image64Helper.ts`を`base64Helper.ts`へリファクタリング (#2080)
* perf: BASE64のデコードを`Buffer.from`から`atob`へ移行

* fix: `imageHelper`to`base64Helper`

* perf: `buffer`に戻す
